Provided Example for S3 Access to Templates
Previous wording was confusing so updated line 16 and provided an example link to a template that the user can deploy. 
Usage here is particularly confusing since the List of Managed Config rules gives identifiers in names-with-dashes form but rule identifiers in the bucket are in ALL_CAPS_WITH_UNDERSCORES form.